requirements.txt
SharedArray
tensorboardX
gast==0.2.2
numba                                                                                                     
matplotlib                                                                                                
fire                                                                                                      
protobuf                                                                                                  
opencv-python                                                                                             
opencv-contrib-python                                                                                     
pybind11                                                                                                                                                                                                                                                                                                             
easydict                                                                                                  
open3d-python                                                                                             
terminaltables                                                                                            
pytest-runner                                                                                             
addict                                                                                                    
pycocotools                                                                                               
imagecorruptions                                                                                          
objgraph                                                                                                  
cachetools
descartes
jupyter
matplotlib
motmetrics<=1.1.3
numpy
pandas>=0.24
Pillow<=6.2.1  # Latest Pillow is incompatible with current torchvision, https://github.com/pytorch/vision/issues/1712
pyquaternion>=0.9.5
scikit-learn
Shapely
tqdm
pyyaml
requests
nuscenes-devkit==1.0.5
